vital_sign/fastapi_server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
pydantic==2.10.3
python-dotenv==1.0.1
python-multipart==0.0.12
pymongo==4.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0
numpy==2.2.0
scikit-learn==1.5.2
pandas==2.2.3